test(training): Run train_step in eager mode for debugging
To diagnose why the CTC loss gradient is not propagating, the @tf.function decorator is commented out from the train_step function.

This forces the training step to run in eager execution mode. While significantly slower, this will definitively determine if the issue lies within TensorFlow's AutoGraph compilation of the complex recognizer model.